redesign Grid class by using free functions and letting the lambda_create_triangulation function do the refinements
