+++ include/exadg/incompressible_navier_stokes/spatial_discretization/operator_dual_splitting.cpp
@@ -133,7 +133,7 @@ OperatorDualSplitting<dim, Number>::initialize_helmholtz_preconditioner()
 
     mg_preconditioner->initialize(this->param.multigrid_data_viscous,
                                   &this->get_dof_handler_u().get_triangulation(),
-                                  this->grid->get_coarse_triangulations(),
+                                  this->grid->coarse_triangulations,
                                   this->get_dof_handler_u().get_fe(),
                                   this->get_mapping(),
                                   this->momentum_operator,
