+++ include/exadg/incompressible_navier_stokes/user_interface/application_base.h
@@ -32,6 +32,7 @@
 // ExaDG
 #include <exadg/convection_diffusion/user_interface/boundary_descriptor.h>
 #include <exadg/grid/grid.h>
+#include <exadg/grid/grid_utilities.h>
 #include <exadg/incompressible_navier_stokes/postprocessor/postprocessor.h>
 #include <exadg/incompressible_navier_stokes/user_interface/boundary_descriptor.h>
 #include <exadg/incompressible_navier_stokes/user_interface/field_functions.h>
